2012-09-26 02:44 UTC+0200 Przemyslaw Czerpak (druzus/at/poczta.onet.pl)   * harbour/include/hbexprb.c   * harbour/include/hbapicdp.h   * harbour/src/rdd/dbfcdx/dbfcdx1.c   * harbour/src/vm/fm.c   * harbour/src/rtl/hbgtcore.c   * harbour/src/rtl/hbsockhb.c   * harbour/contrib/hbct/ctwin.c   * harbour/contrib/hbwin/win_misc.c     * pacified some warnings

+++ harbour/contrib/hbct/ctwin.c
@@ -2177,7 +2177,7 @@ static int hb_ctw_gt_Alert( PHB_GT pGT, PHB_ITEM pMessage, PHB_ITEM pOptions,
                szOptW = hb_arrayGetStrU16( pOptions, i, HB_CDP_ENDIAN_NATIVE, &hOpt, &nLen );
                HB_GTSELF_PUTTEXTW( pGT, iLines + 1, iMnuCol, iClr, szOptW, nLen );
                hb_strfree( hOpt );
-               iMnuCol += nLen + 3;
+               iMnuCol += ( int ) nLen + 3;
             }
             while( HB_GTSELF_DISPCOUNT( pGT ) )
                HB_GTSELF_DISPEND( pGT );
